requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
asttokens==3.0.1
bleach==6.3.0
captum==0.9.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.3
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
Django==6.0.1
django-axes==8.3.1
django-ratelimit==4.1.0
django-simple-history==3.11.0
djangorestframework==3.16.1
executing==2.2.1
filelock==3.28.0
fonttools==4.62.1
fsspec==2026.3.0
grad-cam==1.5.5
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
ImageIO==2.37.3
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kaggle==2.0.1
kagglesdk==0.1.18
kiwisolver==1.5.0
lazy-loader==0.5
lime==0.2.0.1
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
pandas==3.0.2
parso==0.8.6
pillow==11.3.0
platformdirs==4.9.6
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
psycopg2==2.9.11
pure_eval==0.2.3
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.3.0
regex==2026.4.4
reportlab==4.2.5
requests==2.31.0
rich==15.0.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
shap==0.51.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sqlparse==0.5.5
stack-data==0.6.3
stripe==15.0.1
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2026.4.11
tokenizers==0.22.2
torch==2.11.0+cu126
torchaudio==2.11.0+cu126
torchvision==0.26.0+cu126
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.6.1
ttach==0.0.3
typer==0.24.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1